### 2020-12-12 #2903 @ghislainv
@Kevin-Mattheus-Moerman, thank you for handling my submission. I have corrected the DOIs (see ghislainv/forestatrisk@e82b240635446b65b3f598c30302760efbeac752) and will generate a new pdf will `whedon`.
