### 2020-12-16 7c9ffdb
Update paper and forest_country.sh
