### 2020-12-16 3981307
:pencil: Minor changes to paper.md
