+++ paper/paper.md
@@ -40,7 +40,7 @@ link-citations: yes
 
 # Summary
 
-The `ForestAtRisk` Python package can be used to model the spatial probability of deforestation and predict future forest cover in the tropics. The spatial data used to model deforestation comes from georeferenced raster files which can be very large (several gigabytes). The functions available in the `ForestAtRisk` package process large rasters by blocks of data, making calculations fast and efficient. This allows deforestation to be modeled over large geographic areas (eg. at the country scale) and at high resolution (eg $\leq$ 30 m). The `ForestAtRisk` package offers the possibility of using a logistic regression with auto-correlated spatial random effects to model the deforestation process. The spatial random effects make it possible to structure the residual spatial variability in the deforestation process which is not explained by the variables of the model and is often very large. In addition to providing new features, the `ForestAtRisk` Python package is open source (GPLv3 license), cross-platform and scriptable (via Python), user-friendly (functions provided with full documentation and examples), and easily extendable (with statistical models for example). The Python package `ForestAtRisk` was recently used to model deforestation and predict future forest cover by 2100 across the humid tropics.
+The `ForestAtRisk` Python package can be used to model the spatial probability of deforestation and predict future forest cover in the tropics. The spatial data used to model deforestation comes from georeferenced raster files which can be very large (several gigabytes). The functions available in the `ForestAtRisk` package process large rasters by blocks of data, making calculations fast and efficient. This allows deforestation to be modeled over large geographic areas (eg. at the country scale) and at high resolution (eg. $\leq$ 30 m). The `ForestAtRisk` package offers the possibility of using a logistic regression with auto-correlated spatial random effects to model the deforestation process. The spatial random effects make it possible to structure the residual spatial variability in the deforestation process which is not explained by the variables of the model and is often very large. In addition to providing new features, the `ForestAtRisk` Python package is open source (GPLv3 license), cross-platform and scriptable (via Python), user-friendly (functions provided with full documentation and examples), and easily extendable (with statistical models for example). The Python package `ForestAtRisk` was recently used to model deforestation and predict future forest cover by 2100 across the humid tropics.
 
 <!-- The JOSS paper should only include:
 
@@ -64,7 +64,7 @@ docker run --rm \
 
 Commonly called the "Jewels of the Earth", tropical forests shelter 30 million species of plants and animals representing half of the Earth's wildlife and at least two-thirds of its plant species [@Gibson2011]. Through photosynthesis and carbon sequestration, they play an important role in the global carbon cycle, and in regulating the global climate [@Baccini2017]. Despite the many ecosystem services they provide, tropical forests are disappearing at an alarming rate [@Keenan2015; @Vancutsem2020], mostly because of human activities. Currently, around 8 Mha (twice the size of Switzerland) of tropical forest are disappearing each year [@Keenan2015]. Spatial modelling of the deforestation allows identifying the main factors determining the spatial risk of deforestation and quantifying their relative effects. Forecasting forest cover change is paramount as it allows anticipating the consequences of deforestation (in terms of carbon emissions or biodiversity loss) under various technological, political and socio-economic scenarios, and informs decision makers accordingly [@Clark2001]. Because both biodiversity and carbon vary greatly in space [@Allnutt2008; @Baccini2017], it is necessary to provide spatial forecasts of forest cover change to properly quantify biodiversity loss and carbon emissions associated with future deforestation.
 
[…]
