### 2020-12-16 #2903 @ghislainv
Correcting typos in paper's summary.
