### 2020-12-18 d34ce41
Cast defor as float in deforest_diffusion.py
