### 2021-01-04 #2903 @kbarnhart
@ghislainv  Thanks for your submission. I'll now start actively handling it.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

In addition, here are a few minor comments that I would recommend addressing before review begins.
- [ ] As part of the review process, reviewers will need to assess whether it is possible to test the software to verify it has been installed/compiled correctly. I would recommend adding instructions describing how a user might do do this.
- [ ] Reviewers will also assess community guidelines regarding the following tasks: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. I would recommend adding statements for community guidelines to the repo.

If you have any questions about the JOSS review process, or clarifications about the above comments please do not hesitate to reach out to me on this issue or at krbarnhart@usgs.gov
