### 2021-01-11 #2903 @bradyrx
@kbarnhart, this looks like a really cool package. However, I just defended my PhD and got started up last week in a new position at a company. I'll be plenty swamped with onboarding and getting spun up with the team for awhile. I'll have to evaluate after a few months there if I have time to still review at JOSS, but will update you all once I know better there. Best of luck, @ghislainv!
