### 2021-01-11 4be08ec
Modifying the README with 'Installation testing' and 'Contributing' sections
