### 2021-01-11 #2903 @ghislainv
@bradyrx, thanks and good luck with your new position!

@kbarnhart, in response to your first comments, I have included two new sections: "Installation testing" and "Contributing" in the README file of the project. Also, as potential reviewers, I would add @ ecodiv, and @ NikosAlexandris. They are not in the [list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS.
