### 2021-01-15 #2903 @molgor
Hi Kathy,I can do it. It would be my first time to review in the JOSS, so I might bug you with some questions. Best,JuanEl 14/01/2021, a las 13:33, Katy Barnhart <notifications@github.com> escribió:👋 @molgor, @KMarkert, @ethanwhite @ecodiv would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.htmlThe submission is "ForestAtRisk: A Python package for modelling and forecasting deforestation in the tropics".This is a pre-review issue which is used to find reviewers. Once 2-3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks. If you are not able to review and have someone to recommend, please mention them here (when mentioning, please place a space after the @ of a github handle, for example, you would refer to me as "@ kbarnhart").If you are interested in reviewing, I would recommend looking over the journal's conflict of interest policy before the review process starts.If you have any questions about the JOSS review process, please do not hesitate to reach out to me on this issue or at krbarnhart@usgs.gov—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
