### 2021-01-20 027dcbc
Remove zoom in docstring for far.plot.prob
