### 2021-01-20 #2975 @whedon
[REVIEW]: ForestAtRisk: A Python package for modelling and forecasting deforestation in the tropics
**Submitting author:** @ghislainv (<a href="http://orcid.org/0000-0002-1685-4997">Ghislain Vieilledent</a>)
**Repository:** <a href="https://github.com/ghislainv/forestatrisk" target ="_blank">https://github.com/ghislainv/forestatrisk</a>
**Version:** v1.0
