### 2021-02-17 #2975 @molgor
# Checklist for ForestAtRisk repository

  > Yes, the repository can be downloaded without any problem using:
  `git clone` https://github.com/ghislainv/forestatrisk
  > Yes, the repository includes a file with the GNU - GPLv3 in plain text.
  > Yes, according to the Github /insights/ section, the user @ghislainv (i.e. Ghislain Vieilledent ORCID: 0000-0002-1685-4997) is the only contributor and developer of the software,.

  > Yes, the software satisfies the suggested requirements list:
  - [x] Age of software (the software began in 2017)
  - [x] Total lines of code (LOC) (more than 1000 LOC)
  - [x] Whether the software has already been cited in academic papers.
   > It is my understanding that they have a pre-print paper citing it.

- [x] Whether the software is sufficiently useful that it is likely to be cited by your peer group.
  > Certainly, the targeted problem and the statistical methods used are in need by the environmental science community. There are not many Python library capable of implementing CAR processes. The software would allow further developments, analysis and software ensembles.

  > Yes, the instruction given for the conda package manager worked 100% for installing all dependencies and the library itself, However, I did not check the installation with `virtualenv` The installation test performed without any problem using the
  `$forestatrisk` command (as stated in the README.md file).
 >  Yes, I was able to perform all the processes showed in the Getting Started tutorial. All of them explained in the documentation and manuscript.
  > I do not fully undestand this. The software was able to run successfully and without a high impact on my laptop computer. Bayesian approaches (CAR models in particular) are computationally intensive, the software seems to perform well using retail hardware.

  *minor Issue*
  > Yes, this statement is clear, as well as their target audience and applicability. However, this information is only present in the paper section and not in the general documentation. It would be valuable to include this section, either in the main repository (i.e. README.md) or in the full documentation website.
  > yes, instructions brief and clear. The installation is automated and do not require to install dependencies manually.
  > Yes, however the documentation is not easily accessible within the main repository. I recommend to add links to the Jupyter notebook and the documentation in the README.md. A section or subsection would be enough.
  > Yes, complete documentation for all visible functions. The author used best practices for documenting the software. Specifically, the use of Sphinx for extracting and building the web documentation.
  *Issue*
  > Partial, some automated tests are implemented in folder `\test`. However not all of these tests passed. I am appending the results here.

  `======================================== short test summary info ========================================
FAILED test/test_get_started.py::test_model_binomial_iCAR - assert False
FAILED test/test_get_started.py::test_rho - assert False
=============================== 2 failed, 15 passed, 2 warnings in 51.58s ===============================
:100.0%, mean accept. rates= beta:0.350, rho:0.443`

> I'll report this in the ´issues´ section of the repository.

  *Issue*
  > No, links for /Community guidelines/ and a /Code of Conduct/ are mentioned in the last section of the README.md file. However, these links are broken. Please amend this issue.

  > Yes
  > Yes
  > Yes, the author provides a comprehensive list of environmental software for modelling vegetation covers. The author clearly states the limitations of that software and how the proposed project contributes to the field with an innovative, flexible and open source tool.
  > The writing style is within expected academic quality. There were some typos that I suggest in a PR, however, it is not my intention to force the author to acknowledge my suggestions as contributions to the software and they may be dismissed if the author considers so.
  > Yes, all references are appropriately cited and in natbib format (I believe). Also, most of them include their corresponding DOI links.
