### 2021-02-17 b119db3
[JOSS review] suggestions for minor writing edits and typos in the paper.md document.
