### 2021-02-19 94dbef0
Link rendered version of 'Get started' in README
Most users clicking on a 'Get started' link will be looking to quickly see the
commands they can run. Linking to the rendered version instead of the raw
notebook will make this easier for the average user.
