### 2021-02-19 24cb53f
Fix links to contribution and CoC docs
