+++ README.rst
@@ -180,8 +180,8 @@ Contributing
 The ``forestatrisk`` Python package is Open Source and released under
 the `GNU GPL version 3 license <license.html>`__. Anybody who is interested can
 contribute to the package development following our `Community
-guidelines <contributing.html>`__. Every contributor must agree to
-follow the project's `Code of conduct <code_of_conduct.html>`__.
+guidelines <CONTRIBUTING.rst>`__. Every contributor must agree to
+follow the project's `Code of conduct <CODE_OF_CONDUCT.rst>`__.
 
 
 .. |Python version| image:: https://img.shields.io/pypi/pyversions/forestatrisk?logo=python&logoColor=ffd43b&color=306998
