### 2021-02-19 #2975 @ethanwhite
Thanks for the reminder @kbarnhart! I've now completed my review and checked everything excepted `Statement of need` in the `Documentation` section. As @molgor noted this is well covered in the paper but generally absent from the `README` and website. I think a few sentences on both should take care of this.

I checked two boxes that @molgor didn't so I'll briefly note my thinking:
* I considered `Automated tests` satisfied because they are running successfully on GitHub Actions: https://github.com/ghislainv/forestatrisk/actions
* I considered `Community guidelines` satisfied because they are present in the standard named files in the standard location so I knew where to find them even though the links in the README were broken. I submitted a PR to fix the links: https://github.com/ghislainv/forestatrisk/pull/37

@ghislainv - Really nice work! An important package that's been really well developed and documented. It's a pleasure to review good software.
