+++ paper/paper.md
@@ -95,7 +95,7 @@ The Python package `forestatrisk` was recently used to model the spatial probabi
 
 # Figures
 
-![**Map of the spatial probability of deforestation in the Guadeloupe archipelago.** This map has been produced with the `forestatrisk` Python package. Colored pixels represent the extent of the natural old-growth tropical moist forest in 2020. The original map has a 30 m resolution. A relative spatial probability of deforestation was computed for each forest pixel. Probability of deforestation is a function of several explanatory variables describing: topography (altitude and slope), accessibility (distances to nearest road, town, and river), forest landscape (distance to forest edge), deforestation history (distance to past deforestation), and land conservation status (presence of a protected area). This map can be reproduced following the `Get started` tutorial at <https://ecology.ghislainv.fr/forestatrisk>.\label{fig:prob}](prob.png)
+![**Map of the spatial probability of deforestation in the Guadeloupe archipelago.** This map has been produced with the `forestatrisk` Python package. Colored pixels represent the extent of the natural old-growth tropical moist forest in 2020. The original map has a 30 m resolution. A relative spatial probability of deforestation was computed for each forest pixel. Probability of deforestation is a function of several explanatory variables describing: topography (altitude and slope), accessibility (distances to nearest road, town, and river), forest landscape (distance to forest edge), deforestation history (distance to past deforestation), and land conservation status (presence of a protected area). This map can be reproduced following the /Get started/ tutorial at <https://ecology.ghislainv.fr/forestatrisk>.\label{fig:prob}](prob.png)
 
 # Acknowledgements
 
