### 2021-02-21 #2975 @ghislainv
@kbarnhart, @molgor, @ethanwhite: I have answered to your remarks and taken into account your suggestions. In particular, I have added a "Statement of Need" section in the README file of the software and in the Home page of the website.
