### 2021-02-24 #2975 @kbarnhart
Note: I'm going to check off @molgor 's check boxes as all boxes are checked off in [this comment](https://github.com/openjournals/joss-reviews/issues/2975#issuecomment-780272615)
