### 2021-02-24 c501573
Copyedits for JOSS publication
@ghislainv here are some copyedits for your manuscript at JOSS
openjournals/joss-reviews/issues/2975

The changes are pretty minor. If you disagree with any changes, not a problem, let me know and we can discuss.
