### 2021-02-24 #2975 @kbarnhart
@ghislainv I've made a PR with some copyedits and a request for minor revision to Figure 1 at  ghislainv/forestatrisk#40

Once that is merged, I'll do a check that the PDF compiles correctly and we can then move to the final stages of the JOSS process.
