### 2021-02-27 #2975 @ghislainv
@kbarnhart, thanks for proofreading the paper, I will integrate your changes. I am currently modifying the figure of the paper and will keep you informed when it's done.
