### 2021-02-28 fa69245
Edits for JOSS publication
