### 2021-02-28 2a268f6
Correcting format() for Python 2.7 in scalebar.py
