### 2021-02-28 e2ccd20
Replace **dic with update in scalebar.py
