### 2021-02-28 037ee90
Using itertools.chain for multiple **kwargs
