### 2021-03-01 405f4d1
Some final corrections on paper.md
