### 2021-03-01 #2975 @ghislainv
@kbarnhart, I have integrated your suggestions on the text of the paper. I have also modified the figure to include lat/long coordinates and a scale bar.
