### 2021-03-01 #2975 @ghislainv
@kbarnhart, thanks for following up. I have archived version 1.0 of the `forestatrisk` package on Zenodo:
- [x] Make a tagged release of your software, and list the version tag of the archived version here. ➡️ forestatrisk v1.0
- [x] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) ➡️ archived on Zenodo
- [x] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID. ➡️ Metadata (title and authors) are correct.
- [x] Please list the DOI of the archived version here. ➡️ [10.5281/zenodo.4570632](https://doi.org/10.5281/zenodo.4570632)
