### 2021-03-01 #2975 @kbarnhart
@ghislainv I've now recommended that this submission be accepted and published the @openjournals/joss-eics will handle final acceptance. Congratulations.

@molgor and @ethanwhite thank you for contributing reviews to this submission.
