### 2021-03-01 f90ae76
removing some unneeded content that complicates the XML JOSS sends to Crossref
