+++ paper/paper.bib
@@ -246,9 +246,7 @@
   institution = {The Intergovernmental Panel on Climate Change, IPCC},
   year        = {2014},
   editor      = {Pachauri, R. K. and Meyer, L. A.},
-  file        = {IPCC2014_WG1AR5_SPM_FINAL.pdf:IPCC2014_WG1AR5_SPM_FINAL.pdf:PDF;IPCC2014_WG1AR5_Chapter06_FINAL.pdf:IPCC2014_WG1AR5_Chapter06_FINAL.pdf:PDF;IPCC2014_WG1AR5_Chapter09_FINAL.pdf:IPCC2014_WG1AR5_Chapter09_FINAL.pdf:PDF},
   location    = {Geneva, Switzerland},
-  owner       = {ghislain},
   pages       = {3--87},
   timestamp   = {2009.11.25},
   url         = {https://www.ipcc.ch/report/ar5/syr/},
