### 2021-03-01 6cf090b
minor changes in grammar and spelling
