### 2021-03-01 #2975 @danielskatz
@ghislainv - this mostly looks good, but I've suggested some changes in https://github.com/ghislainv/forestatrisk/pull/47 - please merge this or let me know what you disagree with.  (Note that the paper had both UK and US spelling - I've chosen US to be consistent, though UK consistently would also be ok.)
